Rename M650 to M602 and add new views for other m codes

+++ wwwroot/Models/MCodeModel.cs
@@ -4,7 +4,7 @@ namespace gsend.pro.Models
 {
     public class MCodeModel : BaseModel
     {
-        public MCodeModel(BaseModelData modelData, string mCode, string description, int[] seeAlso)
+        public MCodeModel(BaseModelData modelData, string mCode, string description, decimal[] seeAlso)
             : base(modelData)
         {
             MCode = mCode.ToUpper();
@@ -16,6 +16,6 @@ namespace gsend.pro.Models
 
         public string Description { get; }
 
-        public int[] SeeAlso { get; }
+        public decimal[] SeeAlso { get; }
     }
 }
